Refactor Arachne-Publishing
Separation of creation of database entries (books and pages) and copying
files to the cloud.

Publish to arachne-DB has to be done first, because the book-ID generated
by the database, is needed for folder naming in the cloud and the creation
of the TEI-XML. This ID is added to the book's metadata.
